#27 画面サイズを取得する機能をコンポーネント化
